# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QuantSynthica Market API",
    "description": "Unified Financial-Data API integrating yfinance, TradingView-Screener, and Screener.in with quantitative analytics.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/screener/query": {
      "post": {
        "tags": [
          "Screener"
        ],
        "summary": "Execute custom stock screener query",
        "operationId": "query_screener_api_v1_screener_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScreenerQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreenerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/screener/value": {
      "get": {
        "tags": [
          "Screener"
        ],
        "summary": "Predefined screen: Value stocks",
        "operationId": "get_value_screen_api_v1_screener_value_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "market",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Market country",
              "default": "india",
              "title": "Market"
            },
            "description": "Market country"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreenerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/screener/growth": {
      "get": {
        "tags": [
          "Screener"
        ],
        "summary": "Predefined screen: High Growth stocks",
        "operationId": "get_growth_screen_api_v1_screener_growth_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "market",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Market country",
              "default": "india",
              "title": "Market"
            },
            "description": "Market country"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreenerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/screener/momentum": {
      "get": {
        "tags": [
          "Screener"
        ],
        "summary": "Predefined screen: Momentum stocks",
        "operationId": "get_momentum_screen_api_v1_screener_momentum_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "market",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Market country",
              "default": "india",
              "title": "Market"
            },
            "description": "Market country"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreenerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/screener/dividend": {
      "get": {
        "tags": [
          "Screener"
        ],
        "summary": "Predefined screen: High Dividend Yield stocks",
        "operationId": "get_dividend_screen_api_v1_screener_dividend_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "market",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Market country",
              "default": "india",
              "title": "Market"
            },
            "description": "Market country"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreenerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/screener/crypto": {
      "post": {
        "tags": [
          "Screener"
        ],
        "summary": "TradingView Crypto Pairs Screener",
        "operationId": "query_crypto_screener_api_v1_screener_crypto_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenericScreenerRequest",
                "default": {
                  "sort_order": "desc",
                  "limit": 20,
                  "offset": 0
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultiAssetScreenerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/screener/coin": {
      "post": {
        "tags": [
          "Screener"
        ],
        "summary": "TradingView CoinMarketCap Coins Screener",
        "operationId": "query_coin_screener_api_v1_screener_coin_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenericScreenerRequest",
                "default": {
                  "sort_order": "desc",
                  "limit": 20,
                  "offset": 0
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultiAssetScreenerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/screener/forex": {
      "post": {
        "tags": [
          "Screener"
        ],
        "summary": "TradingView Forex Screener",
        "operationId": "query_forex_screener_api_v1_screener_forex_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenericScreenerRequest",
                "default": {
                  "sort_order": "desc",
                  "limit": 20,
                  "offset": 0
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultiAssetScreenerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/screener/futures": {
      "post": {
        "tags": [
          "Screener"
        ],
        "summary": "TradingView Commodity & Index Futures Screener",
        "operationId": "query_futures_screener_api_v1_screener_futures_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenericScreenerRequest",
                "default": {
                  "sort_order": "desc",
                  "limit": 20,
                  "offset": 0
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultiAssetScreenerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/screener/bonds": {
      "post": {
        "tags": [
          "Screener"
        ],
        "summary": "TradingView Sovereign & Corporate Bonds Screener",
        "operationId": "query_bonds_screener_api_v1_screener_bonds_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenericScreenerRequest",
                "default": {
                  "sort_order": "desc",
                  "limit": 20,
                  "offset": 0
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultiAssetScreenerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/screener/cfd": {
      "post": {
        "tags": [
          "Screener"
        ],
        "summary": "TradingView CFD Screener",
        "operationId": "query_cfd_screener_api_v1_screener_cfd_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenericScreenerRequest",
                "default": {
                  "sort_order": "desc",
                  "limit": 20,
                  "offset": 0
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultiAssetScreenerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenericScreenerRequest": {
        "properties": {
          "sort_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort By"
          },
          "sort_order": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Order",
            "default": "desc"
          },
          "limit": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 20
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset",
            "default": 0
          },
          "select_columns": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Select Columns"
          }
        },
        "type": "object",
        "title": "GenericScreenerRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MultiAssetScreenerResponse": {
        "properties": {
          "asset_class": {
            "type": "string",
            "title": "Asset Class"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/MultiAssetItem"
            },
            "type": "array",
            "title": "Items"
          },
          "metadata": {
            "$ref": "#/components/schemas/ResponseMetadata"
          }
        },
        "type": "object",
        "required": [
          "asset_class",
          "total_count",
          "count",
          "items",
          "metadata"
        ],
        "title": "MultiAssetScreenerResponse"
      },
...
      "ResponseMetadata": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Primary provider source, e.g. yfinance, tradingview, screener"
          },
          "retrieved_at": {
            "type": "string",
            "format": "date-time",
            "title": "Retrieved At"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached",
            "description": "Whether this response was served from cache",
            "default": false
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms",
            "description": "Time taken to fetch/compute"
          }
        },
        "type": "object",
        "required": [
          "source"
        ],
        "title": "ResponseMetadata"
      },
...
      "ScreenerFilter": {
        "properties": {
          "market_cap_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap Min"
          },
          "market_cap_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap Max"
          },
          "pe_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pe Min"
          },
          "pe_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pe Max"
          },
          "pb_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pb Max"
          },
          "roe_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roe Min"
          },
          "roce_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roce Min"
          },
          "dividend_yield_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dividend Yield Min"
          },
          "volume_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume Min"
          },
          "rsi_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rsi Min"
          },
          "rsi_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rsi Max"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          }
        },
        "type": "object",
        "title": "ScreenerFilter"
      },
      "ScreenerQueryRequest": {
        "properties": {
          "market": {
            "type": "string",
            "title": "Market",
            "description": "Market/country, e.g. india, america, etc.",
            "default": "india"
          },
          "exchange": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exchange",
            "description": "Exchange filter e.g. NSE, BSE, NASDAQ",
            "default": "NSE"
          },
          "filters": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScreenerFilter"
              },
              {
                "type": "null"
              }
            ]
          },
          "select_columns": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Select Columns",
            "description": "Custom 3000+ TradingView column list (e.g. close|1, MACD.macd)"
          },
          "sort_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort By",
            "description": "Field to sort by",
            "default": "market_cap_basic"
          },
          "sort_order": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Order",
            "description": "'asc' or 'desc'",
            "default": "desc"
          },
          "limit": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 20
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset",
            "default": 0
          }
        },
        "type": "object",
        "title": "ScreenerQueryRequest"
      },
      "ScreenerResponse": {
        "properties": {
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ScreenerStockItem"
            },
            "type": "array",
            "title": "Items"
          },
          "metadata": {
            "$ref": "#/components/schemas/ResponseMetadata"
          }
        },
        "type": "object",
        "required": [
          "total_count",
          "count",
          "items",
          "metadata"
        ],
        "title": "ScreenerResponse"
      },
      "ScreenerStockItem": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "exchange": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exchange"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "change_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Percent"
          },
          "volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          },
          "market_cap": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap"
          },
          "pe_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pe Ratio"
          },
          "pb_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pb Ratio"
          },
          "roe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roe"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "rating": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "extra_fields": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra Fields"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "symbol"
        ],
        "title": "ScreenerStockItem"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}